refactorizar creación de la app para pasar ruta de la bd
